refactor(defaultSetting): defaultSetting으로 systemSetting을 변경 file path를 분리 해 systemSettings에 보관
