+++ src/ProductFactory/parseCSV.js
@@ -1,5 +1,5 @@
 import fs from 'fs';
-import { ERROR_MESSAGES } from '../config/systemSettings.js';
+import { ERROR_MESSAGES } from '../config/defaultSettings.js';
 
 function convertProductValue(header, value) {
   if (value === 'null') return null;
